added task 02 notes taking app
