+++ 03 - php database access/zhenhaoli_task02/NotesTaker/dao/UserDAO.php
@@ -1,6 +1,6 @@
 <?php
 
-require_once './DBConnection.php';
+require_once 'DBConnection.php';
 require_once './Utils.php';
 
 class UserDAO extends DBConnection
