+++ 03 - php database access/zhenhaoli_task02/NotesTaker/register.php
@@ -7,7 +7,7 @@
    4. if fail then register again
 */
 
-require_once './UserDAO.php';
+require_once './dao/UserDAO.php';
 require_once './Utils.php';
 Utils::start_session_onlyif_no_session();
